mapping/requirements.txt
kafka-python
hazelcast-python-client